removed duplicated code from rook, queen and bishop classes, wrote a new function that is involked by the rook, queen and bishop pieces (this new function is in the board/Board.jsx file/dir), did a bit of tidy up (deleted files and dirs that i dont need)
