+++ frontend/frontend/src/board/Board.jsx
@@ -154,3 +154,24 @@ export function pieceToProcess(board, oldPieceLocation, index, pieceType) {
 			break;
 	}
 }
+
+export function legalMoves(board, newX, newY, move1, move2) {
+
+	const legalMoves = [];
+	
+	// i am not using the value of i here, i know LOL.......
+	for (var i = 0; i < board.length; i++) {
+		if (newX >= 0 && newX < Utils.ROWS && newY >= 0 && newY < Utils.COLS) {
+			if (board[newX][newY] === Utils.NULL) {
+				board[newX][newY] = Utils.iCanMoveToHere;
+				legalMoves.push([newX, newY]);
+			} else {
+				break;
+			}
+													
+			newX += move1;
+			newY += move2;
+		}
+	}
+	return legalMoves;
+}
