+++ frontend/frontend/src/king/king.jsx
@@ -11,15 +11,15 @@ class king extends ChessPiece {
 		const col = oldKingLocation[1];
 		
 		const kingMoves = [
-			[-1, 1],
-			[-1, 0],
-			[-1, -1],
-			[0, -1],
-			[0, 1],
-			[1, -1],
-			[1, 0],				
-			[1, 1],
-			[0, 1]
+			[-1, 1], // up, left
+			[-1, 0], // left
+			[-1, -1], // down, left
+			[0, -1], // down
+			[0, 1], // up
+			[1, -1], // down, right
+			[1, 0],	// right			
+			[1, 1], // up, right
+			[0, 1] // up  
 		];
 		
 		const legalMoves = [];
