+++ frontend/frontend/src/main.jsx
@@ -1,6 +1,5 @@
-import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
-//import Board from './components/Board.jsx';
+import { createRoot } from 'react-dom/client';
+import App from './App.jsx';
 
 createRoot(document.getElementById('root')).render(
     <App />
